% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_folder_if_not_exists}
\alias{create_folder_if_not_exists}
\title{create folder function}
\usage{
create_folder_if_not_exists(folder_path)
}
\arguments{
\item{folder_path}{folder directory to create}
}
\description{
create folder function
}
\examples{
create_folder_if_not_exists(folder_path)

}
