% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotContaminationPattern.R
\name{PlotContaminationPattern}
\alias{PlotContaminationPattern}
\title{Plot Contamination Pattern}
\usage{
PlotContaminationPattern(
  obj,
  Outdir,
  Name,
  OverlapRatio = 0.5,
  CELLANNOTATION = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{obj}{Seurat Object return by Step1D}

\item{Outdir}{Output Directory of scCLINIC Results}

\item{Name}{Name of Run}

\item{OverlapRatio}{Step1B: The overlap ratio to carry forward to Step1C, default using 0.5, decrease to have coarse resolution. User manual annotation metadata colname, if CELLANNOTATION = TRUE,}

\item{CELLANNOTATION}{default is FALSE, TRUE if using user cluster annotation}
}
\description{
Plot Contamination Pattern
}
