% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scCLINIC.R
\name{scCLINIC_Main_Function}
\alias{scCLINIC_Main_Function}
\title{This function allows you to run scCLINIC in one step.}
\usage{
scCLINIC_Main_Function(
  Name,
  Input,
  Output,
  resol = 0.8,
  overlapRatioList = c(0.1, 0.25, 0.5, 0.75, 0.9),
  OverlapRatio = 0.5,
  ISThreshold = 0,
  gene_n = 150,
  CELLANNOTATION = FALSE,
  rm_tmp = FALSE,
  SCT = FALSE
)
}
\arguments{
\item{Name}{Name of Run}

\item{Input}{Input Directory of Seurat Object, or seurat object}

\item{resol}{Step1A: FindClusters resolution for initial clustering in Step1, default using Seurat FindClusters default resolution: 0.8}

\item{overlapRatioList}{Step1B: The overlap ratio to try in Step1B (recommended range: 0.1 to 0.5)}

\item{OverlapRatio}{Step1B: The overlap ratio to carry forward to Step1C, default using 0.5, decrease to have coarse resolution. User manual annotation metadata colname, if CELLANNOTATION = TRUE,}

\item{ISThreshold}{Step1C: The threshold for Identity Score (IS) to classify low quality cells}

\item{gene_n}{Number of global markers to used}

\item{CELLANNOTATION}{default is FALSE, TRUE if using user cluster annotation}

\item{rm_tmp}{default is FALSE, TRUE to remove temporary files}

\item{Outdir}{Output Directory of scCLINIC Results}
}
\description{
This function allows you to run scCLINIC in one step.
}
\examples{
scCLINIC_Main_Function("kidney",obj,"~/user/output/")
scCLINIC_Main_Function("kidney","~/user/kidney.rds","~/user/output/")
scCLINIC_Main_Function("kidney","~/user/kidney.rds","~/user/output/", OverlapRatio = "CT.Park",CELLANNOTATION = TRUE)
}
\keyword{all}
\keyword{automated}
\keyword{function}
\keyword{in}
\keyword{one}
\keyword{scCLINIC}
