% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{standard_seurat_clustering}
\alias{standard_seurat_clustering}
\title{scCLINIC subcluster clustering function}
\usage{
standard_seurat_clustering(
  obj,
  res = 0.1,
  npc = 30,
  VF = TRUE,
  Verbose = FALSE,
  SCT = FALSE,
  Global_Top20 = Global_Top20
)
}
\arguments{
\item{obj}{seurat object}

\item{res}{resolution}

\item{npc}{npc, default is 30 (Seurat Default)}

\item{VF}{find variable features, TRUE = enable, FALSE = disable}
}
\description{
scCLINIC subcluster clustering function
}
\examples{
standard_seurat_clustering(obj)

}
\keyword{clustering}
\keyword{function}
\keyword{scCLINIC}
\keyword{subcluster}
