% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STEP1A_GlobalMarkers.R
\name{STEP1A_GlobalMarkers}
\alias{STEP1A_GlobalMarkers}
\title{Step 1A-calculate global markers (before merging)}
\usage{
STEP1A_GlobalMarkers(
  Input,
  Outdir,
  Name,
  resol = 0.8,
  verbose = TRUE,
  SCT = FALSE
)
}
\arguments{
\item{Input}{Input Directory of Seurat Object, or seurat object}

\item{Outdir}{Output Directory of scCLINIC Results}

\item{Name}{Name of Run}

\item{resol}{Step1A: FindClusters resolution for initial clustering in Step1, default using Seurat FindClusters default resolution: 0.8}
}
\description{
Step 1A-calculate global markers (before merging)
}
