% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STEP1B_MergingCluster.R
\name{STEP1B_MergingCluster}
\alias{STEP1B_MergingCluster}
\title{Step1b-merging cluster}
\usage{
STEP1B_MergingCluster(
  obj,
  Outdir,
  Name,
  resol = 0.8,
  overlapRatioList = c(0.1, 0.25, 0.5, 0.75, 0.9),
  gene_n = 150,
  SCT = FALSE
)
}
\arguments{
\item{obj}{Seurat Object return by Step1A}

\item{Outdir}{Output Directory of scCLINIC Results}

\item{Name}{Name of Run}

\item{resol}{Step1A: FindClusters resolution for initial clustering in Step1, default using Seurat FindClusters default resolution: 0.8}

\item{overlapRatioList}{Step1B: The overlap ratio to try in Step1B (recommended range: 0.1 to 0.5)}

\item{gene_n}{Number of global markers to used}

\item{OverlapRatio}{Step1B: The overlap ratio to carry forward to Step1C, default using 0.5, decrease to have coarse resolution. User manual annotation metadata colname, if CELLANNOTATION = TRUE,}
}
\description{
Step1b-merging cluster
}
