% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STEP1D_FilterLowISCluster.R
\name{STEP1D_FilterLowISCluster}
\alias{STEP1D_FilterLowISCluster}
\title{STEP1D Filter Low IS Cluster}
\usage{
STEP1D_FilterLowISCluster(
  obj,
  Outdir,
  Name,
  resol = 0.8,
  OverlapRatio = 0.5,
  ISThreshold = 0,
  CELLANNOTATION = FALSE,
  SCT = FALSE
)
}
\arguments{
\item{obj}{Seurat Object return by Step1C}

\item{Outdir}{Output Directory of scCLINIC Results}

\item{Name}{Name of Run}

\item{resol}{Step1A: FindClusters resolution for initial clustering in Step1, default using Seurat FindClusters default resolution: 0.8}

\item{OverlapRatio}{Step1B: The overlap ratio to carry forward to Step1C, default using 0.5, decrease to have coarse resolution. User manual annotation metadata colname, if CELLANNOTATION = TRUE,}

\item{ISThreshold}{Step1C: The threshold for Indentity Score (IS) to classify low quality cells}

\item{CELLANNOTATION}{default is FALSE, TRUE if using user cluster annotation}
}
\description{
STEP1D Filter Low IS Cluster
}
