% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STEP2A_Subcluster.R
\name{STEP2A_Subcluster}
\alias{STEP2A_Subcluster}
\title{STEP2A Subcluster Major Cluster}
\usage{
STEP2A_Subcluster(
  obj,
  Outdir,
  Name,
  resol = 0.8,
  OverlapRatio = 0.5,
  gene_n = 150,
  CELLANNOTATION = FALSE,
  verbose = FALSE,
  SCT = FALSE
)
}
\arguments{
\item{obj}{Seurat Object return by Step1D}

\item{Outdir}{Output Directory of scCLINIC Results}

\item{Name}{Name of Run}

\item{resol}{Step1A: FindClusters resolution for initial clustering in Step1, default using Seurat FindClusters default resolution: 0.8}

\item{OverlapRatio}{Step1B: The overlap ratio to carry forward to Step1C, default using 0.5, decrease to have coarse resolution. User manual annotation metadata colname, if CELLANNOTATION = TRUE,}

\item{gene_n}{Number of global markers to used}

\item{CELLANNOTATION}{default is FALSE, TRUE if using user cluster annotation}
}
\description{
STEP2A Subcluster Major Cluster
}
