% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Step2B_ContaminationScore.R
\name{STEP2B_ContaminationScore}
\alias{STEP2B_ContaminationScore}
\title{STEP2B Calculate Contamination Score}
\usage{
STEP2B_ContaminationScore(
  obj,
  Outdir,
  Name,
  resol = 0.8,
  OverlapRatio = 0.5,
  gene_n = 150,
  CELLANNOTATION = FALSE,
  SCT = FALSE
)
}
\arguments{
\item{obj}{Seurat Object return by Step1C}

\item{Outdir}{Output Directory of scCLINIC Results}

\item{Name}{Name of Run}

\item{resol}{Step1A: FindClusters resolution for initial clustering in Step1, default using Seurat FindClusters default resolution: 0.8}

\item{OverlapRatio}{Step1B: The overlap ratio to carry forward to Step1C, default using 0.5, decrease to have coarse resolution. User manual annotation metadata colname, if CELLANNOTATION = TRUE,}

\item{gene_n}{Number of global markers to used}

\item{CELLANNOTATION}{default is FALSE, TRUE if using user cluster annotation}
}
\description{
STEP2B Calculate Contamination Score
}
